build: restructured code and added build files for publishin YeaZ to PyPi
