+++ pyproject.toml
@@ -1,0 +1,17 @@
+# pyproject.toml
+[build-system]
+requires = [
+  "setuptools>=45",
+  "wheel",
+  "setuptools_scm[toml]>=6.2"
+]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools_scm]
+write_to = "yeaz/_version.py"
+
+# Ignore DeprecationWarnings
+[tool.pytest.ini_options]
+filterwarnings = [
+    "ignore::DeprecationWarning"
+]
