+++ setup.cfg
@@ -1,0 +1,68 @@
+[metadata]
+name = yeaz
+author = Nicola Dietler and Sahand Jamal Rahi
+author_email = sahand.rahi@epfl.ch
+description = Yeast cells segmentation and tracking with UNet
+keywords =
+    live-cell imaging
+    cell segmentation
+    cell tracking
+    image analysis
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/rahi-lab/YeaZ-GUI
+project_urls =
+    Sahand lab = https://www.epfl.ch/labs/lpbs/professor-rahi/
+classifiers =
+    Development Status :: 3 - Beta
+    Programming Language :: Python :: 3 :: Only
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
+    License :: OSI Approved :: BSD License
+    Intended Audience :: Education
+    Intended Audience :: Science/Research
+    Operating System :: Microsoft :: Windows
+    Operating System :: POSIX
+    Operating System :: Unix
+    Operating System :: MacOS
+    Topic :: Scientific/Engineering
+    Topic :: Scientific/Engineering :: Bio-Informatics
+    Topic :: Scientific/Engineering :: Information Analysis
+    Topic :: Scientific/Engineering :: Image Processing
+    Topic :: Scientific/Engineering :: Visualization
+    Topic :: Utilities
+
+[options]
+packages = find:
+python_requires =
+    >=3.9
+include_package_data = True
+install_requires =
+    numpy>=1.17.4
+    matplotlib>=3.1.1
+    nd2reader>=3.3.0
+    h5py>=3.8.0
+    scikit-image>=0.15.0
+    openpyxl>=3.1.2
+    opencv-python-headless>=4.7.0.72
+    pandas>=0.25.3
+    munkres
+    scikit-learn>=0.0
+    imageio>=2.6.1
+    Pillow>=6.2.1
+    tqdm
+
+[options.extras_require]
+gui = 
+    PyQt6
+
+tensorflow =
+    tensorflow>=2.11.0
+
+torch =
+    torch>=2.0.0
+
+[options.entry_points]
+console_scripts =
+    yeaz = yeaz.__main__:run
