+++ yeaz/disk/image_loader.py
